+++ test/doctests/uix/doc_test.exs
@@ -1,7 +1,7 @@
 defmodule AuroraUixTest.Uix.DocTest do
   use ExUnit.Case, async: true
 
-  doctest AuroraUix.Field
-  doctest AuroraUix.Parser
-  doctest AuroraUix.ResourceConfigUI
+  doctest Aurora.Uix.Field
+  doctest Aurora.Uix.Parser
+  doctest Aurora.Uix.ResourceConfigUI
 end